# Conflicts:
#	code/__DEFINES/DNA.dm
#	code/__DEFINES/bodyparts.dm
#	code/__DEFINES/dcs/signals/signals_atom/signals_atom.dm
#	code/__DEFINES/dcs/signals/signals_mob/signals_living.dm
#	code/__DEFINES/living.dm
#	code/__DEFINES/mobs.dm
#	code/__DEFINES/obj_flags.dm
#	code/__DEFINES/sprite_accessories.dm
#	code/__DEFINES/surgery.dm
#	code/__DEFINES/traits/declarations.dm
#	code/__DEFINES/traits/sources.dm
#	code/__HELPERS/global_lists.dm
#	code/__HELPERS/mobs.dm
#	code/_globalvars/bitfields.dm
#	code/_globalvars/lists/flavor_misc.dm
#	code/_globalvars/lists/objects.dm
#	code/_globalvars/traits/_traits.dm
#	code/datums/brain_damage/phobia.dm
#	code/datums/components/nanites.dm
#	code/datums/diseases/advance/symptoms/organs.dm
#	code/datums/diseases/advance/symptoms/skin.dm
#	code/datums/diseases/advance/symptoms/vision.dm
#	code/datums/dna.dm
#	code/datums/mutable_appearance.dm
#	code/datums/sprite_accessories.dm
#	code/datums/status_effects/debuffs/debuffs.dm
#	code/game/data_huds.dm
#	code/game/objects/effects/anomalies/anomaly_bioscrambler.dm
#	code/game/objects/items/body_egg.dm
#	code/game/objects/items/cosmetics.dm
#	code/game/objects/items/defib.dm
#	code/game/objects/structures/mirror.dm
#	code/game/objects/structures/tables_racks.dm
#	code/modules/antagonists/abductor/equipment/abduction_surgery.dm
#	code/modules/antagonists/changeling/changeling.dm
#	code/modules/antagonists/role_preference/role_latejoins.dm
#	code/modules/antagonists/vampire/datum_vampire.dm
#	code/modules/asset_cache/asset_list_items.dm
#	code/modules/client/preferences/entries/character/species_features/basic.dm
#	code/modules/client/preferences/entries/character/species_features/mutants.dm
#	code/modules/clothing/clothing.dm
#	code/modules/events/disease_outbreak.dm
#	code/modules/mob/dead/observer/observer.dm
#	code/modules/mob/living/brain/brain_item.dm
#	code/modules/mob/living/carbon/alien/special/alien_embryo.dm
#	code/modules/mob/living/carbon/carbon.dm
#	code/modules/mob/living/carbon/carbon_defense.dm
#	code/modules/mob/living/carbon/carbon_update_icons.dm
#	code/modules/mob/living/carbon/human/_species.dm
#	code/modules/mob/living/carbon/human/human.dm
#	code/modules/mob/living/carbon/human/human_update_icons.dm
#	code/modules/mob/living/carbon/human/inventory.dm
#	code/modules/mob/living/carbon/human/species_types/IPC.dm
#	code/modules/mob/living/carbon/human/species_types/abductors.dm
#	code/modules/mob/living/carbon/human/species_types/apid.dm
#	code/modules/mob/living/carbon/human/species_types/diona.dm
#	code/modules/mob/living/carbon/human/species_types/ethereal.dm
#	code/modules/mob/living/carbon/human/species_types/flypeople.dm
#	code/modules/mob/living/carbon/human/species_types/golems.dm
#	code/modules/mob/living/carbon/human/species_types/humans.dm
#	code/modules/mob/living/carbon/human/species_types/jellypeople.dm
#	code/modules/mob/living/carbon/human/species_types/lizardpeople.dm
#	code/modules/mob/living/carbon/human/species_types/monkey.dm
#	code/modules/mob/living/carbon/human/species_types/mothmen.dm
#	code/modules/mob/living/carbon/human/species_types/oozelings.dm
#	code/modules/mob/living/carbon/human/species_types/psyphoza.dm
#	code/modules/mob/living/carbon/human/species_types/pumpkin_man.dm
#	code/modules/mob/living/carbon/human/species_types/shadowpeople.dm
#	code/modules/mob/living/carbon/human/species_types/snail.dm
#	code/modules/mob/living/carbon/human/species_types/zombies.dm
#	code/modules/mob/living/carbon/life.dm
#	code/modules/mob/living/carbon/monkey/monkey.dm
#	code/modules/mob/living/damage_procs.dm
#	code/modules/mob/living/emote.dm
#	code/modules/mob/living/simple_animal/hostile/mining_mobs/hivelord.dm
#	code/modules/reagents/chemistry/reagents/alcohol_reagents.dm
#	code/modules/reagents/chemistry/reagents/other_reagents.dm
#	code/modules/research/designs/limbgrower_designs.dm
#	code/modules/research/xenobiology/crossbreeding/_status_effects.dm
#	code/modules/surgery/bodyparts/_bodyparts.dm
#	code/modules/surgery/bodyparts/dismemberment.dm
#	code/modules/surgery/bodyparts/head.dm
#	code/modules/surgery/bodyparts/parts.dm
#	code/modules/surgery/bodyparts/species_parts/ethereal_bodyparts.dm
#	code/modules/surgery/bodyparts/species_parts/lizard_bodyparts.dm
#	code/modules/surgery/bodyparts/species_parts/misc_bodyparts.dm
#	code/modules/surgery/bodyparts/species_parts/plasmaman_bodyparts.dm
#	code/modules/surgery/bodyparts/species_parts/psyphoza_bodyparts.dm
#	code/modules/surgery/organ_manipulation.dm
#	code/modules/surgery/organs/augments_eyes.dm
#	code/modules/surgery/organs/eyes.dm

+++ code/game/machinery/syndicatebomb.dm
@@ -320,7 +320,8 @@
 	if(adminlog)
 		message_admins(adminlog)
 		log_game(adminlog)
-	qdel(loc)
+	if(isobj(loc))
+		qdel(loc)
 	qdel(src)
 
 // the machine's defusal is mostly done from the wires code, this is here if you want the core itself to do anything.
@@ -454,7 +455,7 @@
 		reactants += G.reagents
 
 	for(var/obj/item/slime_extract/S in beakers)
-		if(S.Uses)
+		if(S.extract_uses)
 			for(var/obj/item/reagent_containers/cup/G in beakers)
 				G.reagents.trans_to(S, G.reagents.total_volume)
 
@@ -541,7 +542,7 @@
 
 /obj/item/syndicatedetonator/attack_self(mob/user)
 	if(timer < world.time)
-		for(var/obj/machinery/syndicatebomb/B in GLOB.machines)
+		for(var/obj/machinery/syndicatebomb/B as anything in SSmachines.get_machines_by_type_and_subtypes(/obj/machinery/syndicatebomb))
 			if(B.active)
 				B.detonation_timer = world.time + BUTTON_DELAY
 				detonated++
